Updated Business Logic exercise, to make it more gradual and easier to fill gaps
Now, there's a template where to put business logic without the need to modify unit tests
