+++ Sample/EventStoreDB/Simple/ECommerce/ShoppingCarts/ShoppingCart.cs
@@ -47,7 +47,7 @@ public record ShoppingCart(
 {
     public bool IsClosed => ShoppingCartStatus.Closed.HasFlag(Status);
 
-    public static ShoppingCart When(ShoppingCart entity, object @event)
+    public static ShoppingCart Evolve(ShoppingCart entity, object @event)
     {
         return @event switch
         {
