+++ Workshops/IntroductionToEventSourcing/07-BusinessLogic/Immutable/PricingCalculator.cs
@@ -1,0 +1,24 @@
+namespace IntroductionToEventSourcing.BusinessLogic.Immutable;
+
+public interface IProductPriceCalculator
+{
+    PricedProductItem Calculate(ProductItem productItems);
+}
+
+public class FakeProductPriceCalculator: IProductPriceCalculator
+{
+    private readonly int value;
+
+    private FakeProductPriceCalculator(int value)
+    {
+        this.value = value;
+    }
+
+    public static FakeProductPriceCalculator Returning(int value) => new(value);
+
+    public PricedProductItem Calculate(ProductItem productItem)
+    {
+        var (productId, quantity) = productItem;
+        return new PricedProductItem(productId, quantity, value);
+    }
+}
