+++ Workshops/IntroductionToEventSourcing/07-BusinessLogic/Immutable/ShoppingCart.cs
@@ -1,10 +1,4 @@
-using System.Text.Json;
-using EventStore.Client;
-using FluentAssertions;
-using IntroductionToEventSourcing.GettingStateFromEvents.Tools;
-using Xunit;
-
-namespace IntroductionToEventSourcing.GettingStateFromEvents.Immutable;
+namespace IntroductionToEventSourcing.BusinessLogic.Immutable;
 using static ShoppingCartEvent;
 
 // EVENTS
@@ -44,7 +38,13 @@ public record PricedProductItem(
     Guid ProductId,
     int Quantity,
     decimal UnitPrice
-);
+)
+{
+    public decimal TotalPrice => Quantity * UnitPrice;
+}
+
+public record ProductItem(Guid ProductId, int Quantity);
+
 
 // ENTITY
 public record ShoppingCart(
@@ -59,7 +59,7 @@ public record ShoppingCart(
     public static ShoppingCart Default() =>
         new (default, default, default, []);
 
-    public static ShoppingCart Evolve(ShoppingCart shoppingCart, object @event)
+    public static ShoppingCart Evolve(ShoppingCart shoppingCart, ShoppingCartEvent @event)
     {
         return @event switch
         {
@@ -119,77 +119,4 @@ public enum ShoppingCartStatus
     Pending = 1,
     Confirmed = 2,
     Canceled = 4
-}
-
-public class GettingStateFromEventsTests: EventStoreDBTest
-{
-    /// <summary>
-    /// Solution - Mutable entity with When method
-    /// </summary>
-    /// <returns></returns>
-    private static async Task<ShoppingCart> GetShoppingCart(EventStoreClient eventStore, string streamName, CancellationToken ct)
-    {
-        var result = eventStore.ReadStreamAsync(
-            Direction.Forwards,
-            streamName,
-            StreamPosition.Start,
-            cancellationToken: ct
-        );
-
-        if (await result.ReadState == ReadState.StreamNotFound)
-            throw new InvalidOperationException("Shopping Cart was not found!");
-
-        return await result
-            .Select(@event =>
-                JsonSerializer.Deserialize(
-                    @event.Event.Data.Span,
-                    Type.GetType(@event.Event.EventType)!
-                )!
-            )
-            .AggregateAsync(
-                ShoppingCart.Default(),
-                ShoppingCart.Evolve,
-                ct
-            );
-    }
-
-    [Fact]
-    public async Task GettingState_FromEventStoreDB_ShouldSucceed()
-    {
-        var shoppingCartId = Guid.NewGuid();
-        var clientId = Guid.NewGuid();
-        var shoesId = Guid.NewGuid();
-        var tShirtId = Guid.NewGuid();
-        var twoPairsOfShoes = new PricedProductItem(shoesId, 2, 100);
-        var pairOfShoes = new PricedProductItem(shoesId, 1, 100);
-        var tShirt = new PricedProductItem(tShirtId, 1, 50);
-
-        var events = new object[]
-        {
-            new ShoppingCartOpened(shoppingCartId, clientId),
-            new ProductItemAddedToShoppingCart(shoppingCartId, twoPairsOfShoes),
-            new ProductItemAddedToShoppingCart(shoppingCartId, tShirt),
-            new ProductItemRemovedFromShoppingCart(shoppingCartId, pairOfShoes),
-            new ShoppingCartConfirmed(shoppingCartId, DateTime.UtcNow),
-            new ShoppingCartCanceled(shoppingCartId, DateTime.UtcNow)
-        };
-
-        var streamName = $"shopping_cart-{shoppingCartId}";
-
-        await AppendEvents(streamName, events, CancellationToken.None);
-
-        var shoppingCart = await GetShoppingCart(EventStore, streamName, CancellationToken.None);
-
-        shoppingCart.Id.Should().Be(shoppingCartId);
-        shoppingCart.ClientId.Should().Be(clientId);
-        shoppingCart.ProductItems.Should().HaveCount(2);
-
-        shoppingCart.ProductItems[0].ProductId.Should().Be(shoesId);
-        shoppingCart.ProductItems[0].Quantity.Should().Be(pairOfShoes.Quantity);
-        shoppingCart.ProductItems[0].UnitPrice.Should().Be(pairOfShoes.UnitPrice);
-
-        shoppingCart.ProductItems[1].ProductId.Should().Be(tShirtId);
-        shoppingCart.ProductItems[1].Quantity.Should().Be(tShirt.Quantity);
-        shoppingCart.ProductItems[1].UnitPrice.Should().Be(tShirt.UnitPrice);
-    }
 }
