+++ Workshops/IntroductionToEventSourcing/07-BusinessLogic/Mutable/BusinessLogicTests.cs
@@ -1,156 +1,20 @@
 using FluentAssertions;
+using IntroductionToEventSourcing.BusinessLogic.Tools;
 using Xunit;
 
 namespace IntroductionToEventSourcing.BusinessLogic.Mutable;
+
 using static ShoppingCartEvent;
-
-// EVENTS
-public abstract record ShoppingCartEvent
-{
-    public record ShoppingCartOpened(
-        Guid ShoppingCartId,
-        Guid ClientId
-    ): ShoppingCartEvent;
-
-    public record ProductItemAddedToShoppingCart(
-        Guid ShoppingCartId,
-        PricedProductItem ProductItem
-    ): ShoppingCartEvent;
-
-    public record ProductItemRemovedFromShoppingCart(
-        Guid ShoppingCartId,
-        PricedProductItem ProductItem
-    ): ShoppingCartEvent;
-
-    public record ShoppingCartConfirmed(
-        Guid ShoppingCartId,
-        DateTime ConfirmedAt
-    ): ShoppingCartEvent;
-
-    public record ShoppingCartCanceled(
-        Guid ShoppingCartId,
-        DateTime CanceledAt
-    ): ShoppingCartEvent;
-
-    // This won't allow external inheritance
-    private ShoppingCartEvent(){}
-}
-
-// VALUE OBJECTS
-public class PricedProductItem
-{
-    public Guid ProductId { get; set; }
-    public decimal UnitPrice { get; set; }
-
-    public int Quantity { get; set; }
-
-    public decimal TotalPrice => Quantity * UnitPrice;
-}
-
-// ENTITY
-public class ShoppingCart
-{
-    public Guid Id { get; private set; }
-    public Guid ClientId { get; private set; }
-    public ShoppingCartStatus Status { get; private set; }
-    public IList<PricedProductItem> ProductItems { get; } = new List<PricedProductItem>();
-    public DateTime? ConfirmedAt { get; private set; }
-    public DateTime? CanceledAt { get; private set; }
-
-    public void Evolve(object @event)
-    {
-        switch (@event)
-        {
-            case ShoppingCartOpened opened:
-                Apply(opened);
-                break;
-            case ProductItemAddedToShoppingCart productItemAdded:
-                Apply(productItemAdded);
-                break;
-            case ProductItemRemovedFromShoppingCart productItemRemoved:
-                Apply(productItemRemoved);
-                break;
-            case ShoppingCartConfirmed confirmed:
-                Apply(confirmed);
-                break;
-            case ShoppingCartCanceled canceled:
-                Apply(canceled);
-                break;
-        }
-    }
-
-    private void Apply(ShoppingCartOpened opened)
-    {
-        Id = opened.ShoppingCartId;
-        ClientId = opened.ClientId;
-        Status = ShoppingCartStatus.Pending;
-    }
-
-    private void Apply(ProductItemAddedToShoppingCart productItemAdded)
-    {
-        var (_, pricedProductItem) = productItemAdded;
-        var productId = pricedProductItem.ProductId;
-        var quantityToAdd = pricedProductItem.Quantity;
-
-        var current = ProductItems.SingleOrDefault(
-            pi => pi.ProductId == productId
-        );
-
-        if (current == null)
-            ProductItems.Add(pricedProductItem);
-        else
-            current.Quantity += quantityToAdd;
-    }
-
-    private void Apply(ProductItemRemovedFromShoppingCart productItemRemoved)
-    {
-        var (_, pricedProductItem) = productItemRemoved;
-        var productId = pricedProductItem.ProductId;
-        var quantityToRemove = pricedProductItem.Quantity;
-
-        var current = ProductItems.Single(
-            pi => pi.ProductId == productId
-        );
-
-        if (current.Quantity == quantityToRemove)
-            ProductItems.Remove(current);
-        else
-            current.Quantity -= quantityToRemove;
-    }
-
-    private void Apply(ShoppingCartConfirmed confirmed)
-    {
-        Status = ShoppingCartStatus.Confirmed;
-        ConfirmedAt = confirmed.ConfirmedAt;
-    }
-
-    private void Apply(ShoppingCartCanceled canceled)
-    {
-        Status = ShoppingCartStatus.Canceled;
-        CanceledAt = canceled.CanceledAt;
-    }
-}
-
-public enum ShoppingCartStatus
-{
-    Pending = 1,
-    Confirmed = 2,
-    Canceled = 4
-}
 
 public static class ShoppingCartExtensions
 {
-    public static ShoppingCart GetShoppingCart(this IEnumerable<object> events)
-    {
-        var shoppingCart = new ShoppingCart();
-
-        foreach (var @event in events)
-        {
-            shoppingCart.Evolve(@event);
-        }
-
-        return shoppingCart;
-    }
+    public static ShoppingCart GetShoppingCart(this EventStore eventStore, Guid shoppingCartId) =>
+        eventStore.ReadStream<ShoppingCartEvent>(shoppingCartId)
+            .Aggregate(ShoppingCart.Initial(), (shoppingCart, @event) =>
+            {
+                shoppingCart.Evolve(@event);
+                return shoppingCart;
+            });
 }
 
 public class BusinessLogicTests
@@ -163,47 +27,73 @@ public class BusinessLogicTests
         var clientId = Guid.NewGuid();
         var shoesId = Guid.NewGuid();
         var tShirtId = Guid.NewGuid();
-        var twoPairsOfShoes =
-            new PricedProductItem
-            {
-                ProductId = shoesId, Quantity = 2, UnitPrice = 100
-            };
-        var pairOfShoes =
-            new PricedProductItem
-            {
-                ProductId = shoesId, Quantity = 1, UnitPrice = 100
-            };
-        var tShirt =
-            new PricedProductItem
-            {
-                ProductId = tShirtId, Quantity = 1, UnitPrice = 50
-            };
+        var twoPairsOfShoes = new ProductItem { ProductId = shoesId, Quantity = 2 };
+        var pairOfShoes = new ProductItem { ProductId = shoesId, Quantity = 1 };
+        var tShirt = new ProductItem { ProductId = tShirtId, Quantity = 1 };
 
-        // TODO: Fill the events object with results of your business logic
-        // to be the same as events below
-        var events = new List<object>
+        var shoesPrice = 100;
+        var tShirtPrice = 50;
+
+        var pricedPairOfShoes = new PricedProductItem { ProductId = shoesId, Quantity = 1, UnitPrice = shoesPrice };
+        var pricedTShirt = new PricedProductItem { ProductId = tShirtId, Quantity = 1, UnitPrice = tShirtPrice };
+
+        var eventStore = new EventStore();
+
+        // Open
+        var shoppingCart = ShoppingCart.Open(shoppingCartId, clientId);
+        eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+
+        // Add Two Pair of Shoes
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        shoppingCart.AddProduct(
+            FakeProductPriceCalculator.Returning(shoesPrice),
+            twoPairsOfShoes
+        );
+        eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+
+        // Add T-Shirt
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        shoppingCart.AddProduct(
+            FakeProductPriceCalculator.Returning(tShirtPrice),
+            tShirt
+        );
+        eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+
+        // Remove a pair of shoes
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        shoppingCart.RemoveProduct(pricedPairOfShoes);
+        eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+
+        // Confirm
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        shoppingCart.Confirm();
+        eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+
+        // Try Cancel
+        var exception = Record.Exception(() =>
         {
-            // new ShoppingCartOpened(shoppingCartId, clientId),
-            // new ProductItemAddedToShoppingCart(shoppingCartId, twoPairsOfShoes),
-            // new ProductItemAddedToShoppingCart(shoppingCartId, tShirt),
-            // new ProductItemRemovedFromShoppingCart(shoppingCartId, pairOfShoes),
-            // new ShoppingCartConfirmed(shoppingCartId, DateTime.UtcNow),
-            // new ShoppingCartCanceled(shoppingCartId, DateTime.UtcNow)
-        };
+            shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+            shoppingCart.Cancel();
+            eventStore.AppendToStream(shoppingCartId, shoppingCart.UncommittedEvents);
+        });
+        exception.Should().BeOfType<InvalidOperationException>();
 
-        var shoppingCart = events.GetShoppingCart();
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
 
         shoppingCart.Id.Should().Be(shoppingCartId);
         shoppingCart.ClientId.Should().Be(clientId);
         shoppingCart.ProductItems.Should().HaveCount(2);
-
-        shoppingCart.ProductItems[0].ProductId.Should().Be(shoesId);
-        shoppingCart.ProductItems[0].Quantity.Should().Be(pairOfShoes.Quantity);
-        shoppingCart.ProductItems[0].UnitPrice.Should().Be(pairOfShoes.UnitPrice);
         shoppingCart.Status.Should().Be(ShoppingCartStatus.Confirmed);
 
-        shoppingCart.ProductItems[1].ProductId.Should().Be(tShirtId);
-        shoppingCart.ProductItems[1].Quantity.Should().Be(tShirt.Quantity);
-        shoppingCart.ProductItems[1].UnitPrice.Should().Be(tShirt.UnitPrice);
+        shoppingCart.ProductItems[0].Should().BeEquivalentTo(pricedPairOfShoes);
+        shoppingCart.ProductItems[1].Should().BeEquivalentTo(pricedTShirt);
+
+        var events = eventStore.ReadStream<ShoppingCartEvent>(shoppingCartId);
+        events.Should().HaveCount(5);
+        events[0].Should().BeOfType<ShoppingCartOpened>();
+        events[1].Should().BeOfType<ProductItemAddedToShoppingCart>();
+        events[2].Should().BeOfType<ProductItemAddedToShoppingCart>();
+        events[3].Should().BeOfType<ProductItemRemovedFromShoppingCart>();
+        events[4].Should().BeOfType<ShoppingCartConfirmed>();
     }
 }
