+++ Workshops/IntroductionToEventSourcing/07-BusinessLogic/Mutable/ShoppingCart.cs
@@ -1,0 +1,175 @@
+namespace IntroductionToEventSourcing.BusinessLogic.Mutable;
+
+using static ShoppingCartEvent;
+
+// EVENTS
+public abstract record ShoppingCartEvent
+{
+    public record ShoppingCartOpened(
+        Guid ShoppingCartId,
+        Guid ClientId
+    ): ShoppingCartEvent;
+
+    public record ProductItemAddedToShoppingCart(
+        Guid ShoppingCartId,
+        PricedProductItem ProductItem
+    ): ShoppingCartEvent;
+
+    public record ProductItemRemovedFromShoppingCart(
+        Guid ShoppingCartId,
+        PricedProductItem ProductItem
+    ): ShoppingCartEvent;
+
+    public record ShoppingCartConfirmed(
+        Guid ShoppingCartId,
+        DateTime ConfirmedAt
+    ): ShoppingCartEvent;
+
+    public record ShoppingCartCanceled(
+        Guid ShoppingCartId,
+        DateTime CanceledAt
+    ): ShoppingCartEvent;
+
+    // This won't allow external inheritance
+    private ShoppingCartEvent() { }
+}
+
+// VALUE OBJECTS
+public class PricedProductItem
+{
+    public Guid ProductId { get; set; }
+    public decimal UnitPrice { get; set; }
+    public int Quantity { get; set; }
+    public decimal TotalPrice => Quantity * UnitPrice;
+}
+
+public class ProductItem
+{
+    public Guid ProductId { get; set; }
+    public int Quantity { get; set; }
+}
+
+// ENTITY
+public class ShoppingCart
+{
+    public Guid Id { get; private set; }
+    public Guid ClientId { get; private set; }
+    public ShoppingCartStatus Status { get; private set; }
+    public IList<PricedProductItem> ProductItems { get; } = new List<PricedProductItem>();
+    public DateTime? ConfirmedAt { get; private set; }
+    public DateTime? CanceledAt { get; private set; }
+
+    public ShoppingCartEvent[] UncommittedEvents { get; private set; } = [];
+
+    public void Evolve(object @event)
+    {
+        switch (@event)
+        {
+            case ShoppingCartOpened opened:
+                Apply(opened);
+                break;
+            case ProductItemAddedToShoppingCart productItemAdded:
+                Apply(productItemAdded);
+                break;
+            case ProductItemRemovedFromShoppingCart productItemRemoved:
+                Apply(productItemRemoved);
+                break;
+            case ShoppingCartConfirmed confirmed:
+                Apply(confirmed);
+                break;
+            case ShoppingCartCanceled canceled:
+                Apply(canceled);
+                break;
+        }
+    }
+
+    public static ShoppingCart Initial() => new();
+
+    public static ShoppingCart Open(
+        Guid cartId,
+        Guid clientId)
+    {
+        throw new NotImplementedException("Fill the implementation part");
+    }
+
+    private void Apply(ShoppingCartOpened opened)
+    {
+        Id = opened.ShoppingCartId;
+        ClientId = opened.ClientId;
+        Status = ShoppingCartStatus.Pending;
+    }
+
+    public void AddProduct(
+        IProductPriceCalculator priceCalculator,
+        ProductItem productItem
+    )
+    {
+        throw new NotImplementedException("Fill the implementation part");
+    }
+
+    private void Apply(ProductItemAddedToShoppingCart productItemAdded)
+    {
+        var (_, pricedProductItem) = productItemAdded;
+        var productId = pricedProductItem.ProductId;
+        var quantityToAdd = pricedProductItem.Quantity;
+
+        var current = ProductItems.SingleOrDefault(
+            pi => pi.ProductId == productId
+        );
+
+        if (current == null)
+            ProductItems.Add(pricedProductItem);
+        else
+            current.Quantity += quantityToAdd;
+    }
+
+    public void RemoveProduct(PricedProductItem productItemToBeRemoved)
+    {
+        throw new NotImplementedException("Fill the implementation part");
+    }
+
+    private void Apply(ProductItemRemovedFromShoppingCart productItemRemoved)
+    {
+        var (_, pricedProductItem) = productItemRemoved;
+        var productId = pricedProductItem.ProductId;
+        var quantityToRemove = pricedProductItem.Quantity;
+
+        var current = ProductItems.Single(
+            pi => pi.ProductId == productId
+        );
+
+        if (current.Quantity == quantityToRemove)
+            ProductItems.Remove(current);
+        else
+            current.Quantity -= quantityToRemove;
+    }
+
+    public void Confirm()
+    {
+        throw new NotImplementedException("Fill the implementation part");
+    }
+
+    private void Apply(ShoppingCartConfirmed confirmed)
+    {
+        Status = ShoppingCartStatus.Confirmed;
+        ConfirmedAt = confirmed.ConfirmedAt;
+    }
+
+    public void Cancel()
+    {
+        throw new NotImplementedException("Fill the implementation part");
+    }
+
+    private void Apply(ShoppingCartCanceled canceled)
+    {
+        Status = ShoppingCartStatus.Canceled;
+        CanceledAt = canceled.CanceledAt;
+    }
+}
+
+public enum ShoppingCartStatus
+{
+    Pending = 1,
+    Confirmed = 2,
+    Canceled = 4
+}
