+++ Workshops/IntroductionToEventSourcing/Solved/07-BusinessLogic/Immutable/ShoppingCartService.cs
@@ -1,0 +1,106 @@
+namespace IntroductionToEventSourcing.BusinessLogic.Immutable;
+using static ShoppingCartEvent;
+using static ShoppingCartCommand;
+
+public abstract record ShoppingCartCommand
+{
+    public record OpenShoppingCart(
+        Guid ShoppingCartId,
+        Guid ClientId
+    ): ShoppingCartCommand;
+
+    public record AddProductItemToShoppingCart(
+        Guid ShoppingCartId,
+        ProductItem ProductItem
+    );
+
+    public record RemoveProductItemFromShoppingCart(
+        Guid ShoppingCartId,
+        PricedProductItem ProductItem
+    );
+
+    public record ConfirmShoppingCart(
+        Guid ShoppingCartId
+    );
+
+    public record CancelShoppingCart(
+        Guid ShoppingCartId
+    ): ShoppingCartCommand;
+
+    private ShoppingCartCommand() {}
+}
+
+public static class ShoppingCartService
+{
+    public static ShoppingCartOpened Handle(OpenShoppingCart command)
+    {
+        var (shoppingCartId, clientId) = command;
+
+        return new ShoppingCartOpened(
+            shoppingCartId,
+            clientId
+        );
+    }
+
+    public static ProductItemAddedToShoppingCart Handle(
+        IProductPriceCalculator priceCalculator,
+        AddProductItemToShoppingCart command,
+        ShoppingCart shoppingCart
+    )
+    {
+        var (cartId, productItem) = command;
+
+        if (shoppingCart.IsClosed)
+            throw new InvalidOperationException(
+                $"Adding product item for cart in '{shoppingCart.Status}' status is not allowed.");
+
+        var pricedProductItem = priceCalculator.Calculate(productItem);
+
+        return new ProductItemAddedToShoppingCart(
+            cartId,
+            pricedProductItem
+        );
+    }
+
+    public static ProductItemRemovedFromShoppingCart Handle(
+        RemoveProductItemFromShoppingCart command,
+        ShoppingCart shoppingCart
+    )
+    {
+        var (cartId, productItem) = command;
+
+        if (shoppingCart.IsClosed)
+            throw new InvalidOperationException(
+                $"Adding product item for cart in '{shoppingCart.Status}' status is not allowed.");
+
+        if (!shoppingCart.HasEnough(productItem))
+            throw new InvalidOperationException("Not enough product items to remove");
+
+        return new ProductItemRemovedFromShoppingCart(
+            cartId,
+            productItem
+        );
+    }
+
+    public static ShoppingCartConfirmed Handle(ConfirmShoppingCart command, ShoppingCart shoppingCart)
+    {
+        if (shoppingCart.IsClosed)
+            throw new InvalidOperationException($"Confirming cart in '{shoppingCart.Status}' status is not allowed.");
+
+        return new ShoppingCartConfirmed(
+            shoppingCart.Id,
+            DateTime.UtcNow
+        );
+    }
+
+    public static ShoppingCartCanceled Handle(CancelShoppingCart command, ShoppingCart shoppingCart)
+    {
+        if (shoppingCart.IsClosed)
+            throw new InvalidOperationException($"Canceling cart in '{shoppingCart.Status}' status is not allowed.");
+
+        return new ShoppingCartCanceled(
+            shoppingCart.Id,
+            DateTime.UtcNow
+        );
+    }
+}
