+++ Workshops/IntroductionToEventSourcing/Solved/07-BusinessLogic/Mutable/Solution2/Aggregate.cs
@@ -1,0 +1,8 @@
+namespace IntroductionToEventSourcing.BusinessLogic.Mutable.Solution2;
+
+public interface IAggregate
+{
+    Guid Id { get; }
+
+    void Evolve(object @event) { }
+}
