+++ Workshops/IntroductionToEventSourcing/Solved/07-BusinessLogic/Mutable/Solution2/BusinessLogicTests.cs
@@ -1,65 +1,20 @@
 using FluentAssertions;
+using IntroductionToEventSourcing.BusinessLogic.Tools;
 using Xunit;
 
 namespace IntroductionToEventSourcing.BusinessLogic.Mutable.Solution2;
 
-// EVENTS
-public abstract record ShoppingCartEvent
-{
-    public record ShoppingCartOpened(
-        Guid ShoppingCartId,
-        Guid ClientId
-    ): ShoppingCartEvent;
-
-    public record ProductItemAddedToShoppingCart(
-        Guid ShoppingCartId,
-        PricedProductItem ProductItem
-    ): ShoppingCartEvent;
-
-    public record ProductItemRemovedFromShoppingCart(
-        Guid ShoppingCartId,
-        PricedProductItem ProductItem
-    ): ShoppingCartEvent;
-
-    public record ShoppingCartConfirmed(
-        Guid ShoppingCartId,
-        DateTime ConfirmedAt
-    ): ShoppingCartEvent;
-
-    public record ShoppingCartCanceled(
-        Guid ShoppingCartId,
-        DateTime CanceledAt
-    ): ShoppingCartEvent;
-
-    // This won't allow external inheritance
-    private ShoppingCartEvent(){}
-}
-
-// VALUE OBJECTS
-public record ProductItem(
-    Guid ProductId,
-    int Quantity
-);
-
-public record PricedProductItem(
-    Guid ProductId,
-    int Quantity,
-    decimal UnitPrice
-);
+using static ShoppingCartEvent;
 
 public static class ShoppingCartExtensions
 {
-    public static ShoppingCart GetShoppingCart(this IEnumerable<object> events)
-    {
-        var shoppingCart = (ShoppingCart)Activator.CreateInstance(typeof(ShoppingCart), true)!;
-
-        foreach (var @event in events)
-        {
-            shoppingCart.Evolve(@event);
-        }
-
-        return shoppingCart;
-    }
+    public static ShoppingCart GetShoppingCart(this EventStore eventStore, Guid shoppingCartId) =>
+        eventStore.ReadStream<ShoppingCartEvent>(shoppingCartId)
+            .Aggregate(ShoppingCart.Initial(), (shoppingCart, @event) =>
+            {
+                shoppingCart.Evolve(@event);
+                return shoppingCart;
+            });
 }
 
 public class BusinessLogicTests
@@ -71,61 +26,73 @@ public class BusinessLogicTests
         var clientId = Guid.NewGuid();
         var shoesId = Guid.NewGuid();
         var tShirtId = Guid.NewGuid();
-
-        var twoPairsOfShoes = new ProductItem(shoesId, 2);
-        var pairOfShoes = new ProductItem(shoesId, 1);
-        var tShirt = new ProductItem(tShirtId, 1);
+        var twoPairsOfShoes = new ProductItem { ProductId = shoesId, Quantity = 2 };
+        var pairOfShoes = new ProductItem { ProductId = shoesId, Quantity = 1 };
+        var tShirt = new ProductItem { ProductId = tShirtId, Quantity = 1 };
 
         var shoesPrice = 100;
         var tShirtPrice = 50;
 
-        var pricedPairOfShoes = new PricedProductItem(shoesId, 1, shoesPrice);
-        var pricedTShirt = new PricedProductItem(tShirtId, 1, tShirtPrice);
+        var pricedPairOfShoes = new PricedProductItem { ProductId = shoesId, Quantity = 1, UnitPrice = shoesPrice };
+        var pricedTShirt = new PricedProductItem { ProductId = tShirtId, Quantity = 1, UnitPrice = tShirtPrice };
 
-        var events = new List<object>();
+        var eventStore = new EventStore();
 
         // Open
-        var (opened, shoppingCart) = ShoppingCart.Open(shoppingCartId, clientId);
-        events.Add(opened);
+        var (opened,_) = ShoppingCart.Open(shoppingCartId, clientId);
+        eventStore.AppendToStream(shoppingCartId, [opened]);
 
-        // Add two pairs of shoes
-        shoppingCart = events.GetShoppingCart();
-        events.Add(shoppingCart.AddProduct(FakeProductPriceCalculator.Returning(shoesPrice), twoPairsOfShoes));
+        // Add Two Pair of Shoes
+        var shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        ShoppingCartEvent result = shoppingCart.AddProduct(
+            FakeProductPriceCalculator.Returning(shoesPrice),
+            twoPairsOfShoes
+        );
+        eventStore.AppendToStream(shoppingCartId, [result]);
 
         // Add T-Shirt
-        shoppingCart = events.GetShoppingCart();
-        events.Add(shoppingCart.AddProduct(FakeProductPriceCalculator.Returning(tShirtPrice), tShirt));
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        result = shoppingCart.AddProduct(
+            FakeProductPriceCalculator.Returning(tShirtPrice),
+            tShirt
+        );
+        eventStore.AppendToStream(shoppingCartId, [result]);
 
-        // Remove pair of shoes
-        events.Add(shoppingCart.RemoveProduct(pricedPairOfShoes));
+        // Remove a pair of shoes
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        result = shoppingCart.RemoveProduct(pricedPairOfShoes);
+        eventStore.AppendToStream(shoppingCartId, [result]);
 
         // Confirm
-        // Uncomment line below and debug to find bug.
-        // shoppingCart = events.GetShoppingCart();
-        events.Add(shoppingCart.Confirm());
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+        result = shoppingCart.Confirm();
+        eventStore.AppendToStream(shoppingCartId, [result]);
 
-        // Cancel
+        // Try Cancel
         var exception = Record.Exception(() =>
-            {
-                events.Add(shoppingCart.Cancel());
-            }
-        );
+        {
+            shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
+            result = shoppingCart.Cancel();
+            eventStore.AppendToStream(shoppingCartId, [result]);
+        });
         exception.Should().BeOfType<InvalidOperationException>();
 
-        // Uncomment line below and debug to find bug.
-        // shoppingCart = events.GetShoppingCart();
+        shoppingCart = eventStore.GetShoppingCart(shoppingCartId);
 
         shoppingCart.Id.Should().Be(shoppingCartId);
         shoppingCart.ClientId.Should().Be(clientId);
         shoppingCart.ProductItems.Should().HaveCount(2);
         shoppingCart.Status.Should().Be(ShoppingCartStatus.Confirmed);
 
-        shoppingCart.ProductItems[0].ProductId.Should().Be(shoesId);
-        shoppingCart.ProductItems[0].Quantity.Should().Be(pricedPairOfShoes.Quantity);
-        shoppingCart.ProductItems[0].UnitPrice.Should().Be(pricedPairOfShoes.UnitPrice);
+        shoppingCart.ProductItems[0].Should().BeEquivalentTo(pricedPairOfShoes);
+        shoppingCart.ProductItems[1].Should().BeEquivalentTo(pricedTShirt);
 
-        shoppingCart.ProductItems[1].ProductId.Should().Be(tShirtId);
-        shoppingCart.ProductItems[1].Quantity.Should().Be(pricedTShirt.Quantity);
-        shoppingCart.ProductItems[1].UnitPrice.Should().Be(pricedTShirt.UnitPrice);
+        var events = eventStore.ReadStream<ShoppingCartEvent>(shoppingCartId);
+        events.Should().HaveCount(5);
+        events[0].Should().BeOfType<ShoppingCartOpened>();
+        events[1].Should().BeOfType<ProductItemAddedToShoppingCart>();
+        events[2].Should().BeOfType<ProductItemAddedToShoppingCart>();
+        events[3].Should().BeOfType<ProductItemRemovedFromShoppingCart>();
+        events[4].Should().BeOfType<ShoppingCartConfirmed>();
     }
 }
