+++ Workshops/IntroductionToEventSourcing/Solved/09-BusinessLogic.EventStoreDB/Immutable/BusinessLogic.cs
@@ -18,7 +18,7 @@ public record ShoppingCart(
 
     public static string StreamName(Guid id) => $"shopping_cart-{id}";
 
-    public static ShoppingCart When(ShoppingCart shoppingCart, object @event)
+    public static ShoppingCart Evolve(ShoppingCart shoppingCart, object @event)
     {
         return @event switch
         {
@@ -50,11 +50,7 @@ public record ShoppingCart(
                 {
                     ProductItems = shoppingCart.ProductItems
                         .Select(pi => pi.ProductId == pricedProductItem.ProductId?
-                            new PricedProductItem(
-                                pi.ProductId,
-                                pi.Quantity - pricedProductItem.Quantity,
-                                pi.UnitPrice
-                            )
+                            pi with { Quantity = pi.Quantity - pricedProductItem.Quantity }
                             :pi
                         )
                         .Where(pi => pi.Quantity > 0)
