+++ Workshops/IntroductionToEventSourcing/Solved/09-BusinessLogic.EventStoreDB/Immutable/BusinessLogicTests.cs
@@ -79,7 +79,7 @@ public class BusinessLogicTests: EventStoreDBTest
 
         // Add two pairs of shoes
         await EventStore.GetAndUpdate(
-            ShoppingCart.When,
+            ShoppingCart.Evolve,
             ShoppingCart.Default(),
             command => ShoppingCart.StreamName(command.ShoppingCartId),
             (command, shoppingCart) =>
@@ -90,7 +90,7 @@ public class BusinessLogicTests: EventStoreDBTest
 
         // Add T-Shirt
         await EventStore.GetAndUpdate(
-            ShoppingCart.When,
+            ShoppingCart.Evolve,
             ShoppingCart.Default(),
             command => ShoppingCart.StreamName(command.ShoppingCartId),
             (command, shoppingCart) =>
@@ -101,7 +101,7 @@ public class BusinessLogicTests: EventStoreDBTest
 
         // Remove pair of shoes
         await EventStore.GetAndUpdate(
-            ShoppingCart.When,
+            ShoppingCart.Evolve,
             ShoppingCart.Default(),
             command => ShoppingCart.StreamName(command.ShoppingCartId),
             RemoveProductItemFromShoppingCart.Handle,
@@ -111,7 +111,7 @@ public class BusinessLogicTests: EventStoreDBTest
 
         // Confirm
         await EventStore.GetAndUpdate(
-            ShoppingCart.When,
+            ShoppingCart.Evolve,
             ShoppingCart.Default(),
             command => ShoppingCart.StreamName(command.ShoppingCartId),
             ConfirmShoppingCart.Handle,
@@ -123,7 +123,7 @@ public class BusinessLogicTests: EventStoreDBTest
         var exception = await Record.ExceptionAsync(async () =>
             {
                 await EventStore.GetAndUpdate(
-                    ShoppingCart.When,
+                    ShoppingCart.Evolve,
                     ShoppingCart.Default(),
                     command => ShoppingCart.StreamName(command.ShoppingCartId),
                     CancelShoppingCart.Handle,
@@ -135,7 +135,7 @@ public class BusinessLogicTests: EventStoreDBTest
         exception.Should().BeOfType<InvalidOperationException>();
 
         var shoppingCart = await EventStore.Get(
-            ShoppingCart.When,
+            ShoppingCart.Evolve,
             ShoppingCart.Default(),
             ShoppingCart.StreamName(shoppingCartId),
             CancellationToken.None
