+++ Workshops/IntroductionToEventSourcing/Solved/11-OptimisticConcurrency.EventStoreDB/Immutable/BusinessLogic.cs
@@ -17,7 +17,7 @@ public record ShoppingCart(
 
     public static string StreamName(Guid id) => $"shopping_cart-{id}";
 
-    public static ShoppingCart When(ShoppingCart shoppingCart, object @event)
+    public static ShoppingCart Evolve(ShoppingCart shoppingCart, object @event)
     {
         return @event switch
         {
@@ -49,11 +49,7 @@ public record ShoppingCart(
                 {
                     ProductItems = shoppingCart.ProductItems
                         .Select(pi => pi.ProductId == pricedProductItem.ProductId?
-                            new PricedProductItem(
-                                pi.ProductId,
-                                pi.Quantity - pricedProductItem.Quantity,
-                                pi.UnitPrice
-                            )
+                            pi with { Quantity = pi.Quantity - pricedProductItem.Quantity }
                             :pi
                         )
                         .Where(pi => pi.Quantity > 0)
